requirements.txt
boto3==1.34.120
botocore==1.34.120
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.0
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
tabulate==0.9.0
urllib3==2.2.1
